A quick type cast fix

+++ src/fighter/donkey/status.rs
@@ -66,7 +66,7 @@ unsafe extern "C" fn donkey_specials_main_loop(fighter: &mut L2CFighterCommon) -
 /// If there are more than (# of DKs in the match * # of barrels allowed per DK)
 /// existing at a time, DK will be unable to pull out a barrel.
 pub unsafe fn barrel_check() -> bool {
-    if smash::app::lua_bind::ItemManager::get_num_of_active_item(*ITEM_KIND_BARREL) >= vl::param_special_s::barrel_count * DK_COUNT {
+    if smash::app::lua_bind::ItemManager::get_num_of_active_item(*ITEM_KIND_BARREL) >= vl::param_special_s::barrel_count as u64 * DK_COUNT {
         return false;
     }
     return true;
